requirements.txt
altgraph==0.17.4
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
chardet==5.2.0
cssselect2==0.8.0
defusedxml==0.7.1
gitdb==4.0.12
GitPython==3.1.45
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pycparser==2.22
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
pywin32-ctypes==0.2.3
smmap==5.0.2
tinycss2==1.4.0
watchdog==6.0.0
webencodings==0.5.1
